+++ spec/ensql_spec.rb
@@ -2,6 +2,7 @@
 require 'ensql'
 require 'ensql/sequel_adapter'
 require 'ensql/active_record_adapter'
+require 'ensql/postgres_adapter'
 
 RSpec.describe Ensql do
 
@@ -39,6 +40,9 @@ RSpec.describe Ensql do
     ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
     Ensql.adapter = Ensql::ActiveRecordAdapter.new
     expect { Ensql.run('select * from not_a_table') }.to raise_error ActiveRecord::StatementInvalid
+
+    Ensql.adapter = Ensql::PostgresAdapter.pool { PG.connect host: 'localhost' }
+    expect { Ensql.run('select * from not_a_table') }.to raise_error PG::UndefinedTable
   end
 
   describe '.sql' do
@@ -71,58 +75,6 @@ RSpec.describe Ensql do
     it 'returns nil' do
       expect(Ensql.run('select 1')).to eq nil
     end
-  end
-
-  describe '.adapter' do
-
-    it 'autodetects Sequel and ActiveRecord' do
-      Ensql.adapter = nil
-      expect(Ensql.adapter).to be_a Ensql::SequelAdapter
-      hide_const 'Sequel'
-      Ensql.adapter = nil
-      expect(Ensql.adapter).to be_a Ensql::ActiveRecordAdapter
-    end
-
-    it 'raises if autodetection fails' do
-      Ensql.adapter = nil
-      hide_const 'Sequel'
-      hide_const 'ActiveRecord'
-      expect { Ensql.adapter }.to raise_error Ensql::Error, including("Couldn't autodetect an adapter")
-    end
-
-    it 'can be manually set' do
-      expect { Ensql.adapter = :foo }
-        .to change { Ensql.adapter }.to :foo
-    end
-
-    it 'warns if using deprecated adapters' do
-      expect { Ensql.adapter = Ensql::ActiveRecordAdapter }.to output(/deprecated/).to_stderr
-      expect { Ensql.adapter = Ensql::SequelAdapter }.to output(/deprecated/).to_stderr
-    end
-
-    describe 'thread-safety' do
-
-      it 'is autodetected in all threads' do
-        Ensql.adapter = nil
-        expect(Ensql.adapter).to eq Thread.new { Ensql.adapter }.join.value
-      end
-
-      it 'is shared from main thread' do
-        Ensql.adapter = :foo
-        Thread.new { expect(Ensql.adapter).to eq :foo }
-      end
-
-      it 'is isolated between child threads' do
-        Ensql.adapter = :foo
-        Thread.new {
-          Ensql.adapter = :bar
-          expect(Ensql.adapter).to eq :bar
-        }.join.value
-        expect(Ensql.adapter).to eq :foo
-      end
-
-    end
-
   end
 
   it "has a version number" do